requirements.txt
fastapi[all]
uvicorn
python-dotenv
PyJWT
pytest
